+++ R/MultiLevelIPF-package.R
@@ -18,4 +18,6 @@
 #' @importFrom stats as.formula
 #' @importFrom utils head
 #' @importFrom forcats fct_inorder
+#' @importFrom utils getFromNamespace
+#' @importFrom wrswoR sample_int_crank
 "_PACKAGE"